Added command shortcut.
Bugfix: corrected range for which edits affect lineNumber.

Refactoring: Moved commands to separate file.

Updated extension categories.

Updated Readme: removed sections, added section and image.
